requirements.txt
# requirements.txt
python-telegram-bot
requests
deep-translator